+++ src/auction_suffix/parser/types.rs
@@ -1,15 +1,18 @@
 use ethers::types::U256;
 
+#[derive(Debug)]
 pub struct AuctionWhitelistItem {
     pub address: String,
     pub allowance: u64,
 }
 
+#[derive(Debug)]
 pub struct ResolverWhitelist {
     pub whitelist: Vec<AuctionWhitelistItem>,
     pub interactions: String,
 }
 
+#[derive(Debug)]
 pub struct InteractionAdditionalInfo {
     pub whitelist: Vec<AuctionWhitelistItem>,
     pub public_resolving_deadline: u64,
@@ -25,22 +28,26 @@ pub struct InteractionFlags {
     pub points_count: u8,
 }
 
+#[derive(Debug)]
 pub struct TakerFeeData {
     pub taker_fee_ratio: U256,
     pub taker_fee_receiver: String,
     pub interactions: String,
 }
 
+#[derive(Debug)]
 pub struct PrivateAuctionDeadline {
     pub deadline: u64,
     pub interactions: String,
 }
 
+#[derive(Debug)]
 pub struct AuctionPoint {
-    pub delay: u32,
+    pub delay: u64,
     pub coefficient: u32,
 }
 
+#[derive(Debug)]
 pub struct ParsedAuctionParams {
     pub interactions: String,
     pub points: Vec<AuctionPoint>,
